feat: add_jwt_claims and get_jwt_claims

+++ flask-sqlalchemy/resources/item.py
@@ -1,6 +1,6 @@
 from flask_restful import Resource, reqparse
 from models.item import ItemModel
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims
 
 
 class Item(Resource):
@@ -41,6 +41,12 @@ class Item(Resource):
 
     @jwt_required
     def delete(self, name):
+        # Check if user is admin
+        claims = get_jwt_claims()
+
+        if not claims['is_admin']:
+            return {'message': 'Admin privillages required'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
